requirements.txt
numpy
pandas
scikit-learn
pyyaml
joblib
dvc
dagshub
